chore: refactor kube service columns
Refactors the Kubernetes service columns using /columns and a props.ts.

Part of #11035.
