--- packages/renderer/src/lib/service/ServiceColumnActions.svelte
@@ -1,11 +1,0 @@
-<script lang="ts">
-import ServiceActions from './ServiceActions.svelte';
-import type { ServiceUI } from './ServiceUI';
-
-interface Props {
-  object: ServiceUI;
-}
-let { object }: Props = $props();
-</script>
-
-<ServiceActions service={object} />
