+++ packages/renderer/src/lib/service/columns/Actions.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+import ServiceActions from '../ServiceActions.svelte';
+import type { Props } from './props';
+
+let { object }: Props = $props();
+</script>
+
+<ServiceActions service={object} />
